New function: set_boost_unit_properties
Building a library/executable for use with Boost.Unit means
that the target sources must be compiled with specific
preprocessor definitions, paths to boost headers and be linked
to the boost_unit_test_framework library. This covers most
use cases, and certainly those used in cetbuildtools.

Wrap this setting in a new set_boost_unit_properties function
to simplify setting up tests for the user. Document function and
add a placeholder test, making this fail to note it requiring
implementation!