New function cet_set_output_directories
Add function cet_set_output_directories to the CetInstallDirs
module to set the default output directories for targets.

Use the selected install directories in the CMAKE_INSTALL_XXX
variables to create build output hierarchies that match the
final install hierarchy.

This helps when writing relocatable packages so that any
relative paths coded in will work in both the build and
install trees.