Provide builting getopt for cet_exec_test
cet_exec_test requires use of GNU getopt which is not available
by default on OS X platforms (and when installed can interfere
with the system version).

Import and include a pure bash implementation of GNU getopt
from:

https://github.com/agriffis/pure-getopt

This is MIT licensed, so compatible with cetbuildtools2.

Drop getopt() function into cet_exec_test script, renaming as
`cet-getopt` to avoid conflicts or confusion. Replace calls
to getopt with calls to cet-getopt.