Use UPS sqlite lib name in FindSQLite
UPS installs of sqlite3 rename the library from "libsqlite3.EXT" to
"libsqlite3_ups.EXT". This causes FindSQLite to fail finding the package
as the new name is unknown.

When in a UPS environment where SQLite has been setup, prepend
"sqlite3_ups" to the list of library names to be searched for. This
is a temporary measure to allow usability of cetbuildtools2 with
UPS installed packages, and only couples to UPS via existence, not
behaviour. Non-UPS usage is unaffected.